reorganize transformer codes
