+++ node/src/main/scala/coop/rchain/node/web/ReportingRoutes.scala
@@ -4,13 +4,85 @@ import cats.effect.Sync
 import cats.implicits._
 import cats.{~>, Applicative}
 import com.google.protobuf.ByteString
+import coop.rchain.models.BlockHash._
 import coop.rchain.casper.ReportingCasper
+import coop.rchain.casper.protocol.ProcessedDeploy
+import coop.rchain.casper.util.rholang.{InternalError, ReplayFailure}
 import coop.rchain.crypto.codec.Base16
-import coop.rchain.models.BlockHash.BlockHash
-import io.circe.syntax._
+import coop.rchain.models.{BindPattern, ListParWithRandom, Par, TaggedContinuation}
+import coop.rchain.models.TaggedContinuation.TaggedCont.{Empty, ParBody, ScalaBodyRef}
+import coop.rchain.rspace.ReportingRspace.ReportingEvent
+import coop.rchain.rspace.ReportingTransformer.ReportingRhoStringTransformer._
 import org.http4s.{HttpRoutes, QueryParamDecoder}
+import coop.rchain.rspace.ReportingTransformer
+import coop.rchain.rholang.interpreter.{PrettyPrinter => RhoPrinter}
+import org.http4s.circe.jsonEncoderOf
 
 object ReportingRoutes {
+
+  val prettyPrinter = RhoPrinter()
+  val transformer =
+    new ReportingEventStringTransformer[Par, BindPattern, ListParWithRandom, TaggedContinuation](
+      serializeC = channel => prettyPrinter.buildString(channel),
+      serializeP = p => p.patterns.map(prettyPrinter.buildString).mkString("[", ";", "]"),
+      serializeA = data => data.pars.map(prettyPrinter.buildString).mkString("[", ";", "]"),
+      serializeK = k =>
+        k.taggedCont match {
+          case ParBody(value)      => prettyPrinter.buildString(value.body)
+          case ScalaBodyRef(value) => s"ScalaBodyRef($value)"
+          case Empty               => "empty"
+        }
+    )
+
+  final case class DeployTrace(deployHash: String, source: String, events: List[RhoEvent])
+  sealed trait Report
+  final case class BlockTracesReport(
+      hash: String,
+      traces: List[DeployTrace]
+  ) extends Report
+  final case class BlockNotFound(hash: String)    extends Report
+  final case class BlockReplayError(hash: String) extends Report
+  final case class BlockReplayFailed(hash: String, msg: String, deployId: Option[String])
+      extends Report
+
+  def transformDeploy[C, P, A, K](transformer: ReportingTransformer[C, P, A, K, RhoEvent])(
+      ipd: ProcessedDeploy,
+      events: Seq[ReportingEvent]
+  ): DeployTrace =
+    DeployTrace(
+      ipd.deploy.sig.base16String,
+      ipd.deploy.data.term,
+      events.map(transformer.transformEvent).toList
+    )
+
+  def transforResult[F[_]: Sync](
+      hash: BlockHash,
+      state: F[Option[
+        Either[ReplayFailure, List[(ProcessedDeploy, Seq[ReportingEvent])]]
+      ]]
+  ): F[Report] =
+    state.map(
+      s =>
+        s match {
+          case None => BlockNotFound(hash.base16String)
+          case Some((Left(internalError @ InternalError(_)))) =>
+            BlockReplayFailed(
+              hash.base16String,
+              internalError.toString,
+              None
+            )
+          case Some(Left(failed)) =>
+            BlockReplayFailed(hash.base16String, failed.toString, None)
+          case Some(Right(deploys)) =>
+            val t = transformDeploy(transformer)(_, _)
+            BlockTracesReport(
+              hash.base16String,
+              deploys.map(Function.tupled(t))
+            )
+          case _ => BlockReplayError(hash.base16String)
+        }
+    )
+
   def service[F[_]: Sync, M[_]: Applicative](
       reportingCasper: ReportingCasper[M]
   )(implicit nt: M ~> F): HttpRoutes[F] = {
@@ -18,7 +90,6 @@ object ReportingRoutes {
     import dsl._
     import io.circe.generic.extras.Configuration
     import io.circe.generic.extras.auto._
-    import org.http4s.circe.CirceEntityEncoder._
 
     implicit val genDevConfig: Configuration =
       Configuration.default
@@ -31,9 +102,11 @@ object ReportingRoutes {
 
     object BlockHashQueryParamMatcher extends QueryParamDecoderMatcher[ByteString]("blockHash")
 
+    implicit val prepareEncoder = jsonEncoderOf[F, Report]
+
     HttpRoutes.of[F] {
       case GET -> Root / "trace" :? BlockHashQueryParamMatcher(hash) =>
-        Ok { nt(reportingCasper.trace(hash).map(_.asJson)) }
+        Ok { transforResult(hash, nt(reportingCasper.trace(hash))) }
     }
   }
 }
